code review: moved new search endpoint into its own controller
